Stabilization II
as I stumbled in front of netio, I got  myself up, and made a simple game using and fixing netio, and now after it was checked, game does work on the server part. The stabilization is near, and the only obstacle is about to be dealed with. our next target is client! Hurray!