Stabilization IV
Once upun a time, netio descended to the realm of mortals, and it brought stabilization to the lands of glorytopia. However, evils have stolen the stabilization. They used code spagettificatoin and hid stabilization in the mot dark parts of the world... Today is the day, when stabilization will finally be graced upon glorytopia. Stabilization was found!